This commit includes:
(1) Updates to L2234.electricity script to (a) read in electricity technology costs from the global energy system so that the default costs are consistent with the core model assumptions, (b) remove erroneous capacity factor columns in some Capital / OMfixed tables, and (c) ensure that a capacity factor was being read in for the battery technology.

(2) A new script (L2247.elecS_tech_costs.R) which creates add-ons to (a) update electricity technology costs for the multiple load segments electricity sector consistent with the Iyer et al. PNNL report, (b) create add-on XML files for the ITC and PTC without embedding those assumptions in the cost files themselves, and (c) create an add-on XML for advanced tech electricity technology costs.