utils.R edited online with Bitbucket to allow more than one chunk call on inputs_of
to allow more than one module/chunk call on inputs_of and outputs_of e.g. inputs_of(c("a", "b")) -> all_data; the issues was R being unable to evaluate || on different lengths i.e. is.null is length 1 but chunks =="" could have been more than one so it confused R. Error in is.null(chunks) || chunks == "" : 'length = 2' in coercion to 'logical(1)'. any(chunks == "") in the commit fixes it. Hard to say if something in the new R changed that is causing it.